Create ANG3LS.yaml
i tried to find out how to have my NFT Collection from getgems.io verified that are stored in tonkeeper and what i read told me to come to here. i added the metadata exactly as it reads when i look at it, pretty much copy and pasted. to me something like this makes you feel better, just seeing that it has been verified and if someone goes to purchase some and uses tonkeeper id like for them to be recognized. id like for it to be legit. i submitted one last night but i edited an old one which had 2 cryptos you denied so that i, guessing might have messed that up for me.  well, thank you for your time, i hope this works out